# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniProyecto",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get users",
                "description": "Get all clients of the auth user",
                "operationId": "getClients",
                "responses": {
                    "200": {
                        "description": "Todos los clientes",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create client",
                "description": "Create and return one client",
                "operationId": "create_client",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "logo",
                                    "phone",
                                    "dni",
                                    "enterprise_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "format": "byte"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "dni": {
                                        "type": "string"
                                    },
                                    "coin": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "enterprise_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "address": "Murcia 24/70 Malecon - Habana",
                                    "logo": "base64....",
                                    "phone": "+1522469335",
                                    "dni": "001458796",
                                    "coin": "USD",
                                    "description": "Hola esto es una descripcion",
                                    "enterprise_id": 4
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/clients/{clientId}": {
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update client",
                "description": "Update and return one client",
                "operationId": "update_client",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "logo",
                                    "phone",
                                    "dni",
                                    "enterprise_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "format": "byte"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "dni": {
                                        "type": "string"
                                    },
                                    "coin": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "enterprise_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "address": "Murcia 24/70 Malecon - Habana",
                                    "logo": "base64....",
                                    "phone": "+1522469335",
                                    "dni": "001458796",
                                    "coin": "USD",
                                    "description": "Hola esto es una descripcion",
                                    "enterprise_id": 4
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete client",
                "description": "Delete one client",
                "operationId": "delete_client",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}